Chess_App/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-chess==1.999
python-dotenv==1.0.0
aiofiles==23.2.1
asyncio
typing-extensions==4.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1